requirements.txt
torch>=1.7.0
torchvision>=0.8.1
opencv-python>=4.1.2
numpy>=1.18.5
matplotlib>=3.2.2
PyYAML>=5.3.1
requests>=2.23.0
pandas>=1.1.4
seaborn>=0.11.0
albumentations>=1.3.0
tqdm>=4.64.0
Pillow>=7.1.2
tensorboard>=2.4.1 
flask>=2.0.0
flask_cors>=3.0.10